Split GitHub Actions summary reporter
