+++ src/Platform/Microsoft.Testing.Extensions.GitHubActionsReport/GitHubActionsSummaryReporter.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using Microsoft.Testing.Extensions.GitHubActionsReport.Resources;
-using Microsoft.Testing.Platform;
 using Microsoft.Testing.Platform.CommandLine;
 using Microsoft.Testing.Platform.Configurations;
 using Microsoft.Testing.Platform.Extensions;
@@ -273,498 +272,6 @@ internal sealed partial class GitHubActionsSummaryReporter :
         return Task.CompletedTask;
     }
 
-    private void ClearFlakyRecords(string uid)
-    {
-        if (!_flakyRecordIndicesByUid.TryGetValue(uid, out List<int>? indices))
-        {
-            return;
-        }
-
-        _flakyRecordIndicesByUid.Remove(uid);
-        foreach (int index in indices)
-        {
-            (string existingUid, string key, TestRecord record) = _records[index];
-            _records[index] = (existingUid, key, new TestRecord(
-                record.DisplayName,
-                record.FullyQualifiedName,
-                record.Kind,
-                record.Duration,
-                isFlaky: false,
-                record.Failure));
-        }
-    }
-
-    public async Task OnTestSessionFinishingAsync(ITestSessionContext testSessionContext)
-    {
-        try
-        {
-            testSessionContext.CancellationToken.ThrowIfCancellationRequested();
-
-            if (!_isEnabled)
-            {
-                return;
-            }
-
-            SummarySnapshot snapshot = BuildSnapshot();
-            string assemblyName = _testApplicationModuleInfo.TryGetAssemblyName() ?? "unknown assembly name";
-            int exitCode = _testApplicationProcessExitCode.GetProcessExitCode();
-            CiCoverageSummaryData coverage = CiCoverageSummary.Create(_testCoverageResult, testSessionContext.SessionUid);
-            CiRunSummaryModule module = CreateModule(snapshot, assemblyName, testSessionContext, coverage);
-            if (_shouldDeferToArtifactPostProcessing()
-                && _configuration.GetTestResultDirectory() is { } resultsDirectory
-                && !RoslynString.IsNullOrWhiteSpace(resultsDirectory))
-            {
-                string fragmentPath = await CiRunSummaryAggregation.WriteFragmentAsync(
-                    resultsDirectory,
-                    GitHubActionsSummaryArtifactPostProcessor.Provider,
-                    GitHubActionsSummaryArtifactPostProcessor.ProviderSlug,
-                    module).ConfigureAwait(false);
-                await _messageBus.PublishAsync(
-                    this,
-                    new SessionFileArtifact(
-                        testSessionContext.SessionUid,
-                        new FileInfo(fragmentPath),
-                        GitHubActionsResources.DisplayName,
-                        GitHubActionsResources.Description,
-                        GitHubActionsSummaryArtifactPostProcessor.FragmentArtifactKind)).ConfigureAwait(false);
-                return;
-            }
-
-            await _historyService.WriteAsync([module], testSessionContext.CancellationToken).ConfigureAwait(false);
-            if (!_isSummaryEnabled)
-            {
-                return;
-            }
-
-            string? path = _environment.GetEnvironmentVariable(StepSummaryEnvironmentVariable);
-            if (RoslynString.IsNullOrWhiteSpace(path))
-            {
-                // Outside a GitHub Actions step (or when summaries are unsupported) there is nowhere to
-                // write. Stay quiet apart from a low-noise trace so local/dev runs don't get a warning.
-                if (_logger.IsEnabled(LogLevel.Trace))
-                {
-                    _logger.LogTrace($"'{StepSummaryEnvironmentVariable}' is not set; skipping job summary.");
-                }
-
-                return;
-            }
-
-            if (_writeOnFailureOnly
-                && !snapshot.Records.Any(static record => record.Kind == TerminalKind.Failed)
-                && !GitHubActionsExitCode.IndicatesFailure(exitCode))
-            {
-                return;
-            }
-
-            // The 1 MiB cap applies to the whole GITHUB_STEP_SUMMARY file, which every test project in the job
-            // appends to, so this reporter degrades in stages as the shared file fills up:
-            //
-            //   below 40%  full section, failures expanded into collapsible diagnostics
-            //   40% - 60%  full section, but failures listed as name and duration only
-            //   above 60%  one-line verdict for the whole project
-            //
-            // Shedding the diagnostics first is what keeps the report useful for longest: the list of which tests
-            // failed is a line per failure, while the diagnostics behind it are kilobytes.
-            //
-            // Both the decision and the write happen inside the writer's lock. Deciding beforehand would let every
-            // project that finishes at the same moment observe the same file length and each render a full
-            // section, so the absolute cap would admit the first few and refuse the rest outright — turning a
-            // report where every project degrades gracefully into one where the first get everything and the rest
-            // get nothing.
-            var writer = new StepSummaryWriter(_fileSystem, path!, _logger, StepSummaryMaxWriteAttempts, StepSummaryRetryDelay);
-            if (!await TryAppendRenderedSummaryAsync(writer, snapshot.Records, assemblyName, exitCode, coverage, testSessionContext).ConfigureAwait(false))
-            {
-                await ReportSectionDroppedAsync(writer, testSessionContext).ConfigureAwait(false);
-            }
-        }
-        catch (OperationCanceledException)
-        {
-            throw;
-        }
-        catch (Exception ex)
-        {
-            _logger.LogUnexpectedException(nameof(OnTestSessionFinishingAsync), ex);
-        }
-    }
-
-    /// <summary>
-    /// What rendering a failure's diagnostics needs, without retaining the test node it came from.
-    /// </summary>
-    internal readonly struct PendingFailure
-    {
-        internal PendingFailure(string fullyQualifiedName, string displayName, string? explanation, Exception? exception, string? declaredFilePath, int declaredLine)
-        {
-            FullyQualifiedName = fullyQualifiedName;
-            DisplayName = displayName;
-            Explanation = explanation;
-            Exception = exception;
-            DeclaredFilePath = declaredFilePath;
-            DeclaredLine = declaredLine;
-        }
-
-        internal string FullyQualifiedName { get; }
-
-        internal string DisplayName { get; }
-
-        internal string? Explanation { get; }
-
-        internal Exception? Exception { get; }
-
-        internal string? DeclaredFilePath { get; }
-
-        internal int DeclaredLine { get; }
-    }
-
-    /// <summary>
-    /// Orders failures the way the rendered summary does, so "the ones that will be shown" is decidable before
-    /// the session ends.
-    /// </summary>
-    /// <remarks>
-    /// The uid is the final tie-break, and it is what makes the order total. Distinct tests can share both names —
-    /// duplicate parameterized cases most obviously — and without it their relative order comes from dictionary
-    /// enumeration in one caller and an unstable sort in the other. Retention and rendering could then disagree
-    /// about which of the tied failures falls inside the first <see cref="MaxFailures"/>, so one would be shown
-    /// stripped of its diagnostics while the other held diagnostics that are never rendered.
-    /// </remarks>
-    private static int CompareForRendering(
-        string leftUid,
-        string leftFullyQualifiedName,
-        string leftDisplayName,
-        string rightUid,
-        string rightFullyQualifiedName,
-        string rightDisplayName)
-    {
-        int result = StringComparer.Ordinal.Compare(leftFullyQualifiedName, rightFullyQualifiedName);
-        if (result != 0)
-        {
-            return result;
-        }
-
-        result = StringComparer.Ordinal.Compare(leftDisplayName, rightDisplayName);
-        return result != 0 ? result : StringComparer.Ordinal.Compare(leftUid, rightUid);
-    }
-
-    /// <summary>
-    /// Records the diagnostics for a test's latest terminal state, keeping retention bounded to
-    /// <paramref name="keep"/> entries.
-    /// </summary>
-    /// <remarks>
-    /// A <see langword="null"/> <paramref name="failure"/> clears whatever was held for the UID. In-process retries
-    /// reuse the same UID, so a test that failed and then recovered has to give its retained slot back — leaving it
-    /// would let a passing test evict the diagnostics of one that is still failing, and that failure would then
-    /// render as a bare line with nothing explaining why its details are missing.
-    /// </remarks>
-    internal static /* for testing */ void ApplyPendingFailure(
-        Dictionary<string, PendingFailure> pending,
-        string uid,
-        PendingFailure? failure,
-        int keep)
-    {
-        if (failure is { } value)
-        {
-            pending[uid] = value;
-            TrimToRenderedFailures(pending, keep);
-        }
-        else
-        {
-            pending.Remove(uid);
-        }
-    }
-
-    /// <summary>
-    /// Drops the entries that sort last once more than <paramref name="keep"/> are held, so retention stays
-    /// bounded however many tests fail.
-    /// </summary>
-    internal static /* for testing */ void TrimToRenderedFailures(Dictionary<string, PendingFailure> pending, int keep)
-    {
-        while (pending.Count > keep)
-        {
-            string? worstKey = null;
-            PendingFailure worst = default;
-            foreach (KeyValuePair<string, PendingFailure> candidate in pending)
-            {
-                if (worstKey is null
-                    || CompareForRendering(candidate.Key, candidate.Value.FullyQualifiedName, candidate.Value.DisplayName, worstKey, worst.FullyQualifiedName, worst.DisplayName) > 0)
-                {
-                    worstKey = candidate.Key;
-                    worst = candidate.Value;
-                }
-            }
-
-            pending.Remove(worstKey!);
-        }
-    }
-
-    /// <summary>
-    /// Takes the recorded tests in a deterministic order and attaches diagnostics to the failures that will
-    /// actually be rendered.
-    /// </summary>
-    /// <remarks>
-    /// The order matters twice. It makes which failures get expanded reproducible — dictionary order is arbitrary,
-    /// so without it a run could expand a different twenty failures each time — and it makes the direct path's
-    /// "first <see cref="MaxFailures"/> encountered" agree with the aggregated path's "first
-    /// <see cref="MaxFailures"/> by name", so exactly the failures that need diagnostics are the ones that pay
-    /// for them.
-    /// </remarks>
-    private SummarySnapshot BuildSnapshot()
-    {
-        List<(string Uid, string Key, TestRecord Record)> entries;
-        Dictionary<string, PendingFailure> pending;
-        lock (_stateLock)
-        {
-            entries = [.. _records];
-            pending = new Dictionary<string, PendingFailure>(_pendingFailures, StringComparer.Ordinal);
-        }
-
-        entries.Sort(static (left, right)
-            => CompareForRendering(
-                left.Key,
-                left.Record.FullyQualifiedName,
-                left.Record.DisplayName,
-                right.Key,
-                right.Record.FullyQualifiedName,
-                right.Record.DisplayName));
-
-        var snapshot = new List<TestRecord>(entries.Count);
-        List<CiRunSummaryHistoryTest>? historyTests = _historyService.IsEnabled
-            ? new(Math.Min(entries.Count, GitHubActionsHistoryStore.MaxTotalSamples))
-            : null;
-        Dictionary<(string TestId, string FullyQualifiedName, string DisplayName), int>? historyIndices =
-            _historyService.IsEnabled ? [] : null;
-        int expanded = 0;
-        foreach ((string Uid, string Key, TestRecord Record) entry in entries)
-        {
-            TestRecord record = entry.Record;
-            if (record.Kind == TerminalKind.Failed
-                && expanded < MaxFailures
-                && pending.TryGetValue(entry.Key, out PendingFailure failure))
-            {
-                expanded++;
-                TestFailureDetails? failureDetails = CaptureFailureDetails(failure);
-                if (failureDetails is not null
-                    && _historyService.TryGetStats(
-                        entry.Uid,
-                        record.FullyQualifiedName,
-                        record.DisplayName,
-                        out GitHubActionsHistoryStats historyStats)
-                    && (historyStats.TotalCount > 0 || historyStats.DurationSampleCount > 0))
-                {
-                    failureDetails = new TestFailureDetails(
-                        GitHubActionsAnnotationReporter.FormatHistoryContext(
-                            failureDetails.Message ?? GitHubActionsResources.NoFailureMessageFallback,
-                            historyStats,
-                            _historyService.HistoryWindowInDays),
-                        failureDetails.ExceptionType,
-                        failureDetails.StackTrace,
-                        failureDetails.FilePath,
-                        failureDetails.LineNumber);
-                }
-
-                record = new TestRecord(
-                    record.DisplayName,
-                    record.FullyQualifiedName,
-                    record.Kind,
-                    record.Duration,
-                    record.IsFlaky,
-                    failureDetails);
-            }
-
-            snapshot.Add(record);
-            if (historyTests is not null && historyIndices is not null)
-            {
-                var historyTest = new CiRunSummaryHistoryTest
-                {
-                    TestId = entry.Uid,
-                    DisplayName = record.DisplayName,
-                    FullyQualifiedName = record.FullyQualifiedName,
-                    Outcome = record.Kind switch
-                    {
-                        TerminalKind.Passed => GitHubActionsHistoryOutcome.Passed,
-                        TerminalKind.Failed => GitHubActionsHistoryOutcome.Failed,
-                        TerminalKind.Skipped => GitHubActionsHistoryOutcome.Skipped,
-                        _ => throw new InvalidOperationException($"Unexpected terminal kind '{record.Kind}'."),
-                    },
-                    DurationTicks = record.Duration.Ticks,
-                    IsFlaky = record.IsFlaky,
-                };
-                (string TestId, string FullyQualifiedName, string DisplayName) identity = (
-                    historyTest.TestId,
-                    historyTest.FullyQualifiedName,
-                    historyTest.DisplayName);
-                if (historyIndices.TryGetValue(identity, out int existingIndex))
-                {
-                    historyTests[existingIndex] = MergeHistoryTest(historyTests[existingIndex], historyTest);
-                }
-                else if (historyTests.Count < GitHubActionsHistoryStore.MaxTotalSamples)
-                {
-                    historyIndices.Add(identity, historyTests.Count);
-                    historyTests.Add(historyTest);
-                }
-            }
-        }
-
-        return new SummarySnapshot(snapshot, historyTests ?? []);
-    }
-
-    private static CiRunSummaryHistoryTest MergeHistoryTest(
-        CiRunSummaryHistoryTest existing,
-        CiRunSummaryHistoryTest current)
-        => new()
-        {
-            TestId = existing.TestId,
-            FullyQualifiedName = existing.FullyQualifiedName,
-            DisplayName = existing.DisplayName,
-            Outcome = existing.Outcome == GitHubActionsHistoryOutcome.Failed
-                || current.Outcome == GitHubActionsHistoryOutcome.Failed
-                    ? GitHubActionsHistoryOutcome.Failed
-                    : existing.Outcome == GitHubActionsHistoryOutcome.Passed
-                        || current.Outcome == GitHubActionsHistoryOutcome.Passed
-                            ? GitHubActionsHistoryOutcome.Passed
-                            : GitHubActionsHistoryOutcome.Skipped,
-            DurationTicks = Math.Max(existing.DurationTicks, current.DurationTicks),
-            IsFlaky = existing.IsFlaky || current.IsFlaky,
-        };
-
-    /// <returns>
-    /// <see langword="false"/> only when the writer refused the content because it would have taken the file past
-    /// GitHub's cap. A write that failed for any other reason has already been reported and returns
-    /// <see langword="true"/>, so the caller does not warn about it twice.
-    /// </returns>
-    private async Task<bool> TryAppendRenderedSummaryAsync(
-        StepSummaryWriter writer,
-        IReadOnlyList<TestRecord> snapshot,
-        string assemblyName,
-        int exitCode,
-        CiCoverageSummaryData coverage,
-        ITestSessionContext testSessionContext)
-    {
-        try
-        {
-            return await writer.AppendRenderedStepSummarySectionAsync(
-                currentLength =>
-                {
-                    var budget = SummaryBudget.ForProject(currentLength);
-                    bool condense = budget.Stage is SummaryStage.Condensed or SummaryStage.Unlisted;
-                    string markdown = condense
-                        ? BuildMinimalMarkdown(snapshot, assemblyName, _targetFrameworkMoniker.Value, exitCode)
-                        : BuildMarkdown(snapshot, assemblyName, _targetFrameworkMoniker.Value, exitCode, coverage, _sections, _includeFailureDetails, budget);
-
-                    // The note is only warranted once whole project sections start disappearing. Dropping the
-                    // expanded diagnostics leaves every project and every failing test still named, which is a
-                    // shortened report rather than an incomplete one, and does not need a warning at the top.
-                    return (markdown, condense);
-                },
-
-                // The project count the note quotes is read from the summary file when the note is written, so it
-                // is passed as a factory rather than a string: only the writer holds the file exclusively, and
-                // only it can count without racing a sibling project.
-                BuildTruncationNotice,
-                testSessionContext.CancellationToken,
-                GitHubActionsFailureDetails.EffectiveStepSummaryLimit).ConfigureAwait(false);
-        }
-        catch (Exception ex) when (ex is not OperationCanceledException)
-        {
-            await ReportWriteFailureAsync(writer.Path, ex, testSessionContext).ConfigureAwait(false);
-            return true;
-        }
-    }
-
-    /// <summary>
-    /// Warns that this project's section did not fit in the shared summary, and makes sure the note explaining
-    /// that the report is incomplete is present.
-    /// </summary>
-    private async Task ReportSectionDroppedAsync(StepSummaryWriter writer, ITestSessionContext testSessionContext)
-    {
-        string overflowWarning = string.Format(
-            CultureInfo.InvariantCulture,
-            GitHubActionsResources.StepSummaryLimitExceededWarning,
-            (writer.GetSummaryLength() ?? 0).ToString(CultureInfo.InvariantCulture),
-            GitHubActionsFailureDetails.EffectiveStepSummaryLimit.ToString(CultureInfo.InvariantCulture));
-
-        if (_logger.IsEnabled(LogLevel.Warning))
-        {
-            _logger.LogWarning(overflowWarning);
-        }
-
-        await _outputDevice.DisplayAsync(this, new WarningMessageOutputDeviceData(overflowWarning), testSessionContext.CancellationToken).ConfigureAwait(false);
-
-        // This project's section is dropped entirely, which is exactly what the note at the top of the summary
-        // describes, so make sure it is there even if this project was not condensed. The note is a few hundred
-        // bytes and dropping the section frees far more room than it takes; if even that does not fit, the writer
-        // refuses it and the summary is genuinely full, where silence is what keeps the rest rendered.
-        await TryAppendNoticeOnlyAsync(writer, testSessionContext).ConfigureAwait(false);
-    }
-
-    /// <summary>
-    /// Writes the top-of-file note on its own, for when this project's section was dropped entirely.
-    /// </summary>
-    private async Task TryAppendNoticeOnlyAsync(StepSummaryWriter writer, ITestSessionContext testSessionContext)
-    {
-        try
-        {
-            await writer.AppendStepSummaryWithLeadingNoticeAsync(
-                string.Empty,
-                BuildTruncationNotice,
-                testSessionContext.CancellationToken,
-                GitHubActionsFailureDetails.EffectiveStepSummaryLimit).ConfigureAwait(false);
-        }
-        catch (Exception ex) when (ex is not OperationCanceledException)
-        {
-            await ReportWriteFailureAsync(writer.Path, ex, testSessionContext).ConfigureAwait(false);
-        }
-    }
-
-    /// <summary>
-    /// Surfaces a failed summary write as a warning: failing to write the summary must not fail the test run.
-    /// </summary>
-    private async Task ReportWriteFailureAsync(string path, Exception ex, ITestSessionContext testSessionContext)
-    {
-        string warning = string.Format(CultureInfo.InvariantCulture, GitHubActionsResources.StepSummaryWriteFailedWarning, path, ex.Message);
-        if (_logger.IsEnabled(LogLevel.Warning))
-        {
-            _logger.LogWarning(warning);
-        }
-
-        await _outputDevice.DisplayAsync(this, new WarningMessageOutputDeviceData(warning), testSessionContext.CancellationToken).ConfigureAwait(false);
-    }
-
-    private CiRunSummaryModule CreateModule(
-        SummarySnapshot snapshot,
-        string assemblyName,
-        ITestSessionContext testSessionContext,
-        CiCoverageSummaryData coverage)
-    {
-        CiRunSummaryModule module = CiRunSummaryAggregation.CreateModule(
-            snapshot.Records,
-            assemblyName,
-            _testApplicationModuleInfo.GetCurrentTestApplicationFullPath(),
-            _targetFrameworkMoniker.Value,
-            RuntimeInformation.ProcessArchitecture.ToString().ToLowerInvariant(),
-            _environment.GetEnvironmentVariable(EnvironmentVariableConstants.TESTINGPLATFORM_DOTNETTEST_EXECUTIONID),
-            testSessionContext.SessionUid.Value,
-            GetAttemptNumber(),
-            _testApplicationProcessExitCode.GetProcessExitCode(),
-            coverage: coverage,
-            writeOnFailureOnly: _writeOnFailureOnly);
-        module.HistoryTests = [.. snapshot.HistoryTests];
-        module.GitHubActionsStepSummaryEnabled = _isSummaryEnabled;
-        module.GitHubActionsHistoryPath = _historyService.HistoryPath;
-        module.GitHubActionsHistoryWindowInDays = _historyService.IsEnabled
-            ? _historyService.HistoryWindowInDays
-            : 0;
-        module.GitHubActionsStepSummarySections = GitHubActionsStepSummarySectionsParser.ToPersistedValues(_sections);
-        return module;
-    }
-
-    private sealed class SummarySnapshot(
-        IReadOnlyList<TestRecord> records,
-        IReadOnlyList<CiRunSummaryHistoryTest> historyTests)
-    {
-        public IReadOnlyList<TestRecord> Records { get; } = records;
-
-        public IReadOnlyList<CiRunSummaryHistoryTest> HistoryTests { get; } = historyTests;
-    }
-
     private int GetAttemptNumber()
         => int.TryParse(
             _environment.GetEnvironmentVariable(EnvironmentVariableConstants.TESTINGPLATFORM_DOTNETTEST_ATTEMPTNUMBER),
@@ -774,81 +281,4 @@ internal sealed partial class GitHubActionsSummaryReporter :
             && attemptNumber > 0
                 ? attemptNumber
                 : 1;
-
-    /// <summary>
-    /// Extracts the explanation and exception a failing state carries, or <see langword="null"/> for a state
-    /// that is not a failure. Separate from <see cref="CaptureFailureDetails"/> so each state shape can be
-    /// covered without standing up a whole reporter.
-    /// </summary>
-    internal static /* for testing */ (string? Explanation, Exception? Exception)? TryGetFailureInfo(TestNodeStateProperty? state)
-        => state switch
-        {
-            FailedTestNodeStateProperty failed => (failed.Explanation, failed.Exception),
-            ErrorTestNodeStateProperty error => (error.Explanation, error.Exception),
-            TimeoutTestNodeStateProperty timeout => (timeout.Explanation, timeout.Exception),
-#pragma warning disable CS0618, MTP0001 // Type or member is obsolete
-            CancelledTestNodeStateProperty cancelled => (cancelled.Explanation, cancelled.Exception),
-#pragma warning restore CS0618, MTP0001 // Type or member is obsolete
-            _ => null,
-        };
-
-    /// <summary>
-    /// Captures the diagnostics of a failing test — explanation/exception message, exception type, stack trace and
-    /// the source location — so the job summary can expand the failure beyond its name.
-    /// </summary>
-    /// <remarks>
-    /// The location is resolved the same way <see cref="GitHubActionsAnnotationReporter"/> resolves it: prefer the
-    /// exception's call site (it pinpoints the failing statement) and fall back to the location the test framework
-    /// reported for the test itself, so frameworks without a usable stack trace still get a location. Values are
-    /// clipped here rather than at render time so an enormous stack trace never reaches the aggregation fragment
-    /// written to disk.
-    /// </remarks>
-    private TestFailureDetails? CaptureFailureDetails(PendingFailure failure)
-    {
-        Exception? exception = failure.Exception;
-        string repoRoot = GitHubActionsRepositoryRoot.Resolve(_environment) ?? string.Empty;
-        (string RelativeNormalizedPath, int LineNumber)? stackLocation = StackTraceSourceLocationResolver.TryResolve(
-            exception?.StackTrace,
-            repoRoot,
-            _fileSystem,
-            _logger,
-            StackTraceSourceLocationResolver.SkipAssertionFramesForCurrentRuntime);
-        GitHubActionsSourceLocation? location = stackLocation is { } resolved
-            ? new GitHubActionsSourceLocation(resolved.RelativeNormalizedPath, resolved.LineNumber)
-            : ResolveDeclaredLocation(failure, repoRoot);
-
-        // Fall back on whitespace, not just null: Clip treats a whitespace-only value as absent, so an empty
-        // explanation would otherwise discard a perfectly good exception message.
-        string? explanation = RoslynString.IsNullOrWhiteSpace(failure.Explanation)
-            ? exception?.Message
-            : failure.Explanation;
-
-        return new TestFailureDetails(
-            GitHubActionsFailureDetails.Clip(explanation, GitHubActionsFailureDetails.MaxMessageLength, GitHubActionsFailureDetails.MaxMessageRows),
-            exception?.GetType().FullName,
-            GitHubActionsFailureDetails.Clip(exception?.StackTrace, GitHubActionsFailureDetails.MaxStackTraceLength, GitHubActionsFailureDetails.MaxStackTraceRows),
-            location?.RelativeNormalizedPath,
-            location?.LineNumber ?? 0);
-    }
-
-    /// <summary>
-    /// Resolves the location the test framework declared for the test, from the file and line read out of its
-    /// property bag at capture time. Mirrors <see cref="GitHubActionsAnnotationReporter.TryResolveDeclaredLocation"/>,
-    /// but works from the two small values kept rather than from the node, which is not retained.
-    /// </summary>
-    private GitHubActionsSourceLocation? ResolveDeclaredLocation(PendingFailure failure, string repoRoot)
-    {
-        if (RoslynString.IsNullOrWhiteSpace(failure.DeclaredFilePath))
-        {
-            return null;
-        }
-
-        string? relativeNormalizedPath = StackTraceSourceLocationResolver.TryMakeWorkspaceRelative(failure.DeclaredFilePath!, repoRoot, _fileSystem);
-
-        // A framework that knows the file but not the line reports a sentinel (-1) or 0; GitHub accepts a
-        // 'file'-only annotation, so drop the line rather than emitting an invalid one.
-        return relativeNormalizedPath is null
-            ? null
-            : new GitHubActionsSourceLocation(relativeNormalizedPath, failure.DeclaredLine > 0 ? failure.DeclaredLine : 0);
-    }
 }
